requirements.txt
mysql-connector-python==8.0.31
pandas==2.0.1
python-dotenv==1.0.0